+++ tests/Unit/Response/EmptyResponse.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Aceproxies\ResellerApi\Tests\Unit\Response;
+
+final class EmptyResponse
+{
+}